Cleaned up the code
And it seems to work.  Need to reverse the path list as it currently
outputs the path backwards.  Output seems to be limiting itself to
about 8 steps and then showing (..|..).  Don’t know if this is a
problem with my code or just a SWI-Prolog i/o foible.  Need to limit
the output to one path as the code currently sends the system into a
loop.